fix: rec / med / risk route read available

+++ src/modules/checklist/services/risk-data/upsert-risk-data/upsert-risk.service.ts
@@ -8,19 +8,24 @@ export class UpsertRiskDataService {
   constructor(private readonly riskDataRepository: RiskDataRepository) {}
 
   async execute(upsertRiskDataDto: UpsertRiskDataDto) {
+    const keepEmpty = upsertRiskDataDto.keepEmpty;
+    delete upsertRiskDataDto.keepEmpty;
+
     const riskData = await this.riskDataRepository.upsert(upsertRiskDataDto);
 
-    const isEmpty =
-      riskData.adms.length === 0 &&
-      riskData.recs.length === 0 &&
-      riskData.engs.length === 0 &&
-      riskData.epis.length === 0 &&
-      riskData.generateSources.length === 0 &&
-      !riskData.probability;
+    if (!keepEmpty) {
+      const isEmpty =
+        riskData.adms.length === 0 &&
+        riskData.recs.length === 0 &&
+        riskData.engs.length === 0 &&
+        riskData.epis.length === 0 &&
+        riskData.generateSources.length === 0 &&
+        !riskData.probability;
 
-    if (isEmpty) {
-      await this.riskDataRepository.deleteById(riskData.id);
-      return riskData.id;
+      if (isEmpty) {
+        await this.riskDataRepository.deleteById(riskData.id);
+        return riskData.id;
+      }
     }
 
     return riskData;
